fix resume
